refactor(cogs): Consolidate monitoring cogs into Runner and CommandHandler
Refactored the cogs to improve separation of concerns and reduce redundancy.

- Created Runner cog for background processing and health checks.
- Created CommandHandler cog for all user-facing commands.
- Deleted obsolete monitor, monitoring, and config cogs.
